new printTable; add getIP; add main

+++ mp1/src/structs/structs.go
@@ -3,6 +3,6 @@ package structs
 //Membership  exported
 type Membership struct {
 	ID        string
-	HeartBeat float32
-	LocalTime float32
+	HeartBeat int64
+	LocalTime int64
 }
